Refactor #noissue : 파일 명 및 변수 명 오타 수정
- 파일 명 및 변수 명 오타 수정
- isVaild -> isValid로 오타 수정
