+++ server/src/util/isValidToken.ts
